requirements.txt
streamlit==1.31.0
anthropic==0.18.1
sentence-transformers==2.3.1
psycopg2-binary==2.9.9
tiktoken==0.6.0
numpy==1.26.4
pdfplumber==0.10.3
pgvector==0.2.5
